+++ pyfakefs/tests/fake_filesystem_vs_real_test.py
@@ -22,7 +22,7 @@ import tempfile
 import time
 import unittest
 
-from pyfakefs import fake_filesystem
+from pyfakefs import fake_filesystem, fake_os, fake_open
 from pyfakefs.helpers import IS_PYPY
 
 
@@ -112,8 +112,8 @@ class FakeFilesystemVsRealTest(TestCase):
 
         self.fake_filesystem = fake_filesystem.FakeFilesystem()
         self.fake_filesystem.create_dir(self.fake_base)
-        self.fake_os = fake_filesystem.FakeOsModule(self.fake_filesystem)
-        self.fake_open = fake_filesystem.FakeFileOpen(self.fake_filesystem)
+        self.fake_os = fake_os.FakeOsModule(self.fake_filesystem)
+        self.fake_open = fake_open.FakeFileOpen(self.fake_filesystem)
         self._created_files = []
 
         os.chdir(self.real_base)